requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.11.1
openai==2.5.0
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.12.3
pydantic_core==2.41.4
pymongo==4.11.3
python-dotenv==1.1.1
python-telegram-bot==22.0
rsa==4.9
sniffio==1.3.1
Telethon==1.39.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
